+++ core/src/main/scala/com/raphtory/python/PyRaphtory.scala
@@ -6,6 +6,7 @@ import com.monovore.decline._
 import com.monovore.decline.effect._
 import com.raphtory.internals.management.Py4JServer
 import com.raphtory.internals.management.PythonInterop
+import com.raphtory.internals.management.PythonInterop.disableReflectWarning
 import com.typesafe.scalalogging.Logger
 import org.slf4j.LoggerFactory
 import sun.misc.Unsafe
@@ -21,19 +22,6 @@ object PyRaphtory
         ) {
 
   private val logger: Logger = Logger(LoggerFactory.getLogger(this.getClass))
-
-  def disableReflectWarning(): Unit =
-    try {
-      val theUnsafe = classOf[Unsafe].getDeclaredField("theUnsafe")
-      theUnsafe.setAccessible(true)
-      val u         = theUnsafe.get(null).asInstanceOf[Unsafe]
-      val cls       = Class.forName("jdk.internal.module.IllegalAccessLogger")
-      val logger    = cls.getDeclaredField("logger")
-      u.putObjectVolatile(cls, u.staticFieldOffset(logger), null)
-    }
-    catch {
-      case e: Exception => println(e.getStackTrace.mkString("Array(", ", ", ")"))
-    }
 
   def checkParent(parentID: Long): IO[Unit] =
     IO.blocking(ProcessHandle.current().parent().toScala match {
